Ubah tipe data track_record & ubah logic bukuacara

+++ database/migrations/2024_07_30_132649_create_atlets_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->date('umur');
             $table->enum('jenis_kelamin', ['Pria','Wanita']);
-            $table->double('track_record', 8, 2)->nullable();
+            $table->decimal('track_record', 8, 2)->nullable();
             $table->string('dokumen')->nullable();
             $table->timestamps();
         });
